Decouple ObjectiveRelation from CoreData
Introducing...CoreDataRelation!

(We really ought to start thinking about prefixes as we leave Category
Land.)

At the moment, it works on its own with NSArray. We should consider
supporting NSSet and NSOrderedSet, too, adding categories to them, and
and maybe extracting the idea of ObjectiveRelation completely into a
separate "query-builder" pod, which ObjectiveRecord could depend on and
import just the core it needs.

Moving forward, we could make these interfaces a bit more private, and
we can explore how to better share logic (inheritance feels slightly
heavy-handed, and the fact that ObjectiveRelation has an _objects ivar
that is never utilized down the chain feels...off), but I think it's a
step in the right direction, or a step on the way to the right
direction, at the very least.

Signed-off-by: Stephen Celis <stephen@stephencelis.com>